+++ lib/API/Search.php
@@ -56,7 +56,7 @@ class Search
         $header['total_rows'] 	= $total_rows ?: (int) self::getTotal($sql, $values);
         $header['total_pages'] 	= ceil($header['total_rows']/$records_per_page);
         $header['table'] 		= $tb;
-        $header['stripped_table'] = str_replace(PREFIX, null, $tb);
+        $header['stripped_table'] = str_replace(PREFIX, '', $tb);
         $header['table_label'] 	= self::$cfg->get("tables.$tb.label");
         $header['page'] 		= ($page > $header['total_pages']) ? $header['total_pages'] : $page;