+++ lib/Config/ToFiles.php
@@ -17,7 +17,7 @@ class ToFiles
         self::write_in_file($path2cfg . '/app_data.json', $cfg['main']);
         self::writeTables($path2cfg . '/tables.json', $cfg['tables']);
         foreach ($cfg['tables'] as $tb => $tb_data) {
-            self::writeFields($path2cfg . '/' . \str_replace($prefix, null, $tb) . '.json', $cfg['tables'][$tb]['fields']);
+            self::writeFields($path2cfg . '/' . \str_replace($prefix, '', $tb) . '.json', $cfg['tables'][$tb]['fields']);
         }
     }