+++ modules/geoface/geoface.php
@@ -32,7 +32,7 @@ class geoface_ctrl extends Controller
 
 			$record->persist();
 			// TODO: $new_id is undefined
-			$this->response('ok_insert_geodata', 'success', null, [ 'id' => $new_id ] );
+			$this->response('ok_insert_geodata', 'success', null, [ 'id' => $record->getCore('id') ] );
 
 
 		} catch (\Throwable $e) {