+++ CHANGELOG.md
@@ -4,8 +4,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [4.0.8] - 2021-05-11
+### Added
+- Added template method `print.geodata` as shorthand for `print.plg('geodata')`
+
 ### Fixed
 - Fixed bug with Geodata not being shown
+- Code cleaning
 
 ## [4.0.7] - 2021-04-17
 ### Added